backend/requirements.txt
# Core Web Framework
fastapi==0.110.1
uvicorn==0.25.0

# Database Drivers
aiomysql>=0.2.0
PyMySQL>=1.1.0

# Security & Authentication
python-jose>=3.3.0
pyjwt>=2.10.1
bcrypt==4.1.3
passlib>=1.7.4
cryptography>=42.0.8

# Environment & Configuration
python-dotenv>=1.0.1
pydantic>=2.6.4
email-validator>=2.2.0

# AWS & Cloud Services
boto3>=1.34.129
requests-oauthlib>=2.0.0

# Data Processing & Analysis
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.12.0
python-multipart>=0.0.9
python-dateutil>=2.8.2
openpyxl>=3.1.2
xlrd>=2.0.1

# ============================================================================
# PHASE 10.1: AI/ML INFRASTRUCTURE
# ============================================================================

# Redis for Caching & Queue
redis>=5.0.0
aioredis>=2.0.1

# Celery for Background Tasks
celery>=5.3.4
celery[redis]>=5.3.4
flower>=2.0.1

# Machine Learning Core
scikit-learn>=1.4.0

# Embeddings & Similarity Search
sentence-transformers>=2.3.1
faiss-cpu>=1.7.4

# LLM Integration (via Emergent LLM Key)
openai>=1.10.0
anthropic>=0.18.0
google-generativeai>=0.3.2

# Text Processing
nltk>=3.8.1
spacy>=3.7.0

# ============================================================================
# ADVANCED FEATURES DEPENDENCIES
# ============================================================================

# QR Code & Image Processing
qrcode>=7.4.2
Pillow>=10.2.0
reportlab>=4.0.0

# Email Service
sendgrid>=6.11.0

# ============================================================================
# MONITORING & LOGGING
# ============================================================================

# Metrics & Monitoring
prometheus-client>=0.19.0
python-json-logger>=2.0.7
sentry-sdk>=1.40.0

# ============================================================================
# DEVELOPMENT & TESTING
# ============================================================================

pytest>=8.0.0
black>=24.1.1
isort>=5.13.2
flake8>=7.0.0
mypy>=1.8.0
requests>=2.31.0
jq>=1.6.0
typer>=0.9.0
tzdata>=2024.2
